array of scripts appears to be working
