+++ src/workflows/renderFrames.ts
@@ -1,6 +1,6 @@
 import {proxyActivities} from '@temporalio/workflow';
 import * as activities from '../activities';
-import {EventScript} from '../events/scripts';
+import {EventScript} from '../events/run-scripts';
 import {ScriptConfig} from '../interfaces';
 
 interface Params {
